requirements/test.txt
mock
pytest
pytest-django
pytest-pythonpath
pytest-cov
pdbpp
django-debug-toolbar